[kernel] move std libc related to rtklibc.h
